+++ client/src/ContactInformation/SecondaryContactInformation/SecondaryContactInformationForm.jsx
@@ -26,10 +26,11 @@ class SecondaryContactInformationForm extends Component {
       errors.phoneNumber = "Please enter a valid phone number.";
     if (validator.isEmpty(formValues.phoneNumber + ""))
       errors.phoneNumber = "Phone number is required.";
-    if (!validator.isEmail(formValues.email + ""))
+    if (
+      !validator.isEmpty(this.state.email + "") &&
+      !validator.isEmail(formValues.email + "")
+    )
       errors.email = "Please enter a valid email address.";
-    if (validator.isEmpty(formValues.email + ""))
-      errors.email = "Email address is required.";
     return errors;
   };
 
@@ -102,7 +103,7 @@ class SecondaryContactInformationForm extends Component {
           />
           <Input
             name="email"
-            label="Email"
+            label="Email Address (optional)"
             type="input"
             onChange={this.handleChange}
             wasValidated={this.state.wasValidated}